# gpt-bridge/openapi.yaml
openapi: 3.1.0
info:
  title: HyperFrames GPT Bridge
  version: 1.0.0
  description: Safe HTTP bridge exposing selected HyperFrames Student Kit capabilities to a GPT action or skill.
servers:
  - url: https://YOUR-HOST.example.com
security:
  - bearerAuth: []
paths:
...
  /v1/styles:
    get:
      operationId: searchStyleCards
      summary: Search the HyperFrames style-card registry
      parameters:
        - in: query
          name: query
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200': { description: Matching cards }
...
  /v1/projects:
    get:
      operationId: listProjects
      summary: List video projects
      responses:
        '200': { description: Project list }
    post:
      operationId: createProject
      summary: Create a project from the included starter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project]
              properties:
                project:
                  type: string
                  pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
      responses:
        '201': { description: Created }
...
  /v1/doctor:
    post:
      operationId: runDoctor
      summary: Check the HyperFrames runtime environment
      responses:
        '200': { description: Environment healthy }
        '422': { description: Environment problems found }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...